Move aux core patch player interrupts to own class
