+++ firmware/src/core_a7/aux_core_main.cc
@@ -1,17 +1,14 @@
+#include "aux_core_player.hh"
 #include "conf/hsem_conf.hh"
 #include "core_a7/a7_shared_memory.hh"
 #include "core_a7/async_thread_control.hh"
-#include "core_a7/smp_api.hh"
 #include "debug.hh"
 #include "drivers/hsem.hh"
-#include "drivers/smp.hh"
-#include "drivers/timekeeper.hh"
 #include "dynload/plugin_manager.hh"
 #include "fs/filesystem.hh"
 #include "fs/norflash_layout.hh"
 #include "gui/ui.hh"
 #include "internal_plugin_manager.hh"
-#include "patch_play/patch_player.hh"
 
 using FrameBufferT =
 	std::array<lv_color_t, MetaModule::ScreenBufferConf::width * MetaModule::ScreenBufferConf::height / 4>;
@@ -31,29 +28,22 @@ extern "C" void aux_core_main() {
 
 	pr_info("A7 Core 2 starting\n");
 
-	auto patch_player = A7SharedMemoryS::ptrs.patch_player;
-	auto patch_playloader = A7SharedMemoryS::ptrs.patch_playloader;
-	auto file_storage_proxy = A7SharedMemoryS::ptrs.patch_storage;
-	auto open_patch_manager = A7SharedMemoryS::ptrs.open_patch_manager;
-	auto sync_params = A7SharedMemoryS::ptrs.sync_params;
-	auto patch_mod_queue = A7SharedMemoryS::ptrs.patch_mod_queue;
-	auto ramdisk_storage = A7SharedMemoryS::ptrs.ramdrive;
 #ifdef CONSOLE_USE_USB
 	UartLog::use_usb(A7SharedMemoryS::ptrs.console_buffer);
 #endif
 
 	LVGLDriver gui{MMDisplay::flush_to_screen, MMDisplay::read_input, MMDisplay::wait_cb, framebuf1, framebuf2};
 
-	RamDiskOps ramdisk_ops{*ramdisk_storage};
+	RamDiskOps ramdisk_ops{*A7SharedMemoryS::ptrs.ramdrive};
 	FatFileIO ramdisk{&ramdisk_ops, Volume::RamDisk};
 	AssetFS asset_fs{AssetVolFlashOffset};
 	Filesystem::Init(ramdisk);
-	PluginManager plugin_manager{*file_storage_proxy, ramdisk};
-	Ui ui{*patch_playloader,
-		  *file_storage_proxy,
-		  *open_patch_manager,
-		  *sync_params,
-		  *patch_mod_queue,
+	PluginManager plugin_manager{*A7SharedMemoryS::ptrs.patch_storage, ramdisk};
+	Ui ui{*A7SharedMemoryS::ptrs.patch_playloader,
+		  *A7SharedMemoryS::ptrs.patch_storage,
+		  *A7SharedMemoryS::ptrs.open_patch_manager,
+		  *A7SharedMemoryS::ptrs.sync_params,
+		  *A7SharedMemoryS::ptrs.patch_mod_queue,
 		  plugin_manager,
 		  ramdisk};
 	ui.update_screen();
@@ -61,62 +51,7 @@ extern "C" void aux_core_main() {
 
 	InternalPluginManager internal_plugin_manager{ramdisk, asset_fs};
 
-	struct AuxCoreModulesToRun {
-		uint32_t starting_idx = 1;
-		uint32_t num_modules = 0;
-		uint32_t idx_increment = 2;
-	} modules_to_run;
-
-	constexpr auto PlayModuleListIRQn = SMPControl::IRQn(SMPCommand::PlayModuleList);
-	InterruptManager::register_and_start_isr(PlayModuleListIRQn, 1, 0, [&modules_to_run, &patch_player]() {
-		// Debug::Pin1::high();
-		for (unsigned i = modules_to_run.starting_idx; i < modules_to_run.num_modules;
-			 i += modules_to_run.idx_increment)
-		{
-			patch_player->modules[i]->update();
-		}
-		// Debug::Pin1::low();
-		SMPThread::signal_done();
-	});
-
-	constexpr auto NewModuleListIRQn = SMPControl::IRQn(SMPCommand::NewModuleList);
-	InterruptManager::register_and_start_isr(NewModuleListIRQn, 0, 0, [&modules_to_run]() {
-		modules_to_run.starting_idx = SMPControl::read<SMPRegister::ModuleID>();
-		modules_to_run.num_modules = SMPControl::read<SMPRegister::NumModulesInPatch>();
-		modules_to_run.idx_increment = SMPControl::read<SMPRegister::UpdateModuleOffset>();
-		SMPThread::signal_done();
-	});
-
-	constexpr auto ReadPatchLightsIRQn = SMPControl::IRQn(SMPCommand::ReadPatchLights);
-	InterruptManager::register_and_start_isr(ReadPatchLightsIRQn, 2, 0, [patch_player, &ui]() {
-		if (ui.new_patch_data == false) {
-
-			for (auto &w : ui.lights().watch_lights) {
-				if (w.is_active()) {
-					auto val = patch_player->get_module_light(w.module_id, w.light_id);
-					w.value = val;
-				}
-			}
-
-			for (auto &d : ui.displays().watch_displays) {
-				if (d.is_active()) {
-					auto text = std::span<char>(d.text._data, d.text.capacity);
-					auto sz = patch_player->get_display_text(d.module_id, d.light_id, text);
-					d.text._data[sz] = '\0';
-				}
-			}
-
-			for (auto &p : ui.watched_params().active_watched_params()) {
-				if (p.is_active()) {
-					p.value = patch_player->get_param(p.module_id, p.param_id);
-				}
-			}
-
-			ui.new_patch_data = true;
-		}
-
-		SMPThread::signal_done();
-	});
+	AuxPlayer aux_player{*A7SharedMemoryS::ptrs.patch_player, ui};
 
 	// Wait for M4 to be ready (so USB and SD are available)
 	while (mdrivlib::HWSemaphore<M4CoreReady>::is_locked())
